+++ CHANGELOG.rst
@@ -4,12 +4,15 @@ Changelog
 
 Version 0.1.5
 -------------
+- Fixes wrong sanity check for measurement types and addition of spurious NaN or empty columns in density dataframes [`BBPP82-630`_]
+- Addresses the troubles caused by duplicate region names in brain regions hierarchy [`BBPP82-630`_]
+- Integrate the computation of mtype volumetric densities by Yann Roussel (BBP) [`NSETM-1574`_]
 - Uses sphinx's autodoc and sphinx-click to generated HTML documentation of the command line interface [`NSETM-1484`_]
 - Makes a clearer separation between the splitting of layer 2/3 in terms of hierarchy (json) and the splitting the annotated AIBS volume (nrrd) [`NSETM-1513`_]
 
 Version 0.1.4 (2021-08-12)
 --------------------------
-- migration from gerrit to gitlab: `https://bbpgitlab.epfl.ch/nse/atlas-building-tools` [`NSETM-1562`_]
+- Migration from gerrit to gitlab: `https://bbpgitlab.epfl.ch/nse/atlas-building-tools` [`NSETM-1562`_]
 - Uses the volume slicer of `nse/cgal-pybind` to split layer 2/3 of the AIBS mouse isocortex [`NSETM-1513`_]
 - Creates and re-uses a common atlas option group [`NSETM-1513`_]
 - Makes average inhibitory neuron densities consistent and create refined volumetric density files [`NSETM-1506`_]
@@ -61,6 +64,8 @@ Version 0.1.0 (2021-04-27)
 - CLI to compute a flat map based on the streamlines of a laminar brain region
 - CLI to turn the excel measurement compilation gaba_papers.xlsx of D. Rodarie into a CSV file.
 
+.. _`NSETM-1574`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1574
+.. _`BBPP82-630`: https://bbpteam.epfl.ch/project/issues/browse/BBPP82-630
 .. _`NSETM-1484`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1484
 .. _`NSETM-1562`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1562
 .. _`NSETM-1513`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1513